Decoupled JSend response interceptor
